sesion20/requirements.txt
django
djangorestframework
django-cors-headers
django-http-referrer-policy
